requirements.txt
boto==2.4.1
django-storages==1.1.4
